requirements.txt
pandas
numpy
pycrypto
psq==0.7.0
cryptography
pyjwt
curlify